The @chainable Macro

Chainable.@chainable — Macro

@chainable

A macro that makes functions pipeline-friendly by generating method definitions that work seamlessly with Julia's |> operator. For any function f(x, args...), creates a factory method f(args...) that returns a function x -> f(x, args...), enabling natural use in pipelines.

This supports both types of function assignment syntax:

@chainable function f(x, y) ... end
@chainable f(x, y) = ...

Note that in cases where there are default positional arguments, in order to avoid method dispatch conflicts, the factory method will only accept keyword arguments. This is to ensure that the factory method can be called with no arguments, which would otherwise be ambiguous. For example, the following function:

@chainable LeakyRelu(x, alpha=0.01)::Real = max.(x, alpha*x)

By default, julia will create the methods LeakyRelu(x, alpha) and LeakyRelu(x). The method created by @chainable will then be LeakyRelu(; alpha)::Function. This allows both the syntax 1 |> LeakyRelu() and 1 |> LeakyRelu(alpha = 0.02) to work, but not 1 |> LeakyRelu(0.02) because the right-hand side of the pipe operator now calls the default LeakyRelu(x) method.

Examples

# Make a function pipeline-friendly
@chainable function transform(data, factor, offset)
    return data * factor + offset
end

# Use in a pipeline
1 |> transform(2, 3)  # Same as: transform(1, 2, 3)

# Chain multiple operations
[1, 2, 3] |> 
    transform(2, 0) |>   # multiply by 2
    transform(1, 1)      # add 1

# Works with any number of arguments
@chainable function process(x, a, b, c)
    return x + a + b + c
end

10 |> process(1, 2, 3)  # Returns 16

@chainable increment(x, inc = 1) = x + inc
1 |> increment() # works
1 |> increment(2) # this is undefined behaviour. In this case it fails because increment(x::Int) is defined.
1 |> increment(inc = 2) # works